+++ packages/ng/libraries/user/src/lib/index.ts
@@ -2,6 +2,7 @@ export * from './display/index';
 export * from './tile/index';
 export * from './picture/index';
 export * from './select/index';
+export * from './service/index';
 
 export * from './user.model';
 export * from './user.module';
