+++ packages/ng/libraries/user/src/lib/select/searcher/index.ts
@@ -1,4 +1,2 @@
 export * from './user-searcher.component';
-export * from './user-searcher.model';
 export * from './user-searcher.module';
-export * from './user-searcher.service';
