+++ packages/ng/libraries/user/src/lib/service/index.ts
@@ -1,0 +1,2 @@
+export * from './user-service.model';
+export * from './user-v3.service';
